Externalized Secert and SecretRetriever to the runtime package
