+++ runtime/secret.go
@@ -1,3 +1,8 @@
 package runtime
 
-
+type Secret struct {
+	Identifier string
+	EnvName    string
+	Version    string
+	Path       string
+}
